Organize structure of folders Add folder for DataAccess - interfaces to interact with DB
