+++ Server-Module/Server/Database/DataAccess/DataAccess.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace Server.Database.DataAccess
+{
+    public class DataAccess
+    {
+    }
+}
