+++ Server-Module/Server/Exceptions/NotOwnerException.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Server.Database.Exceptions
+namespace Server.Exceptions
 {
     public class NotOwnerException:Exception
     {
